ref: add routingBuilder and routeFactory
